Rather barebones attempt at transforming reagents into xgm gasses and making humanoids breathe them. This is NOT intended to be merged in this state at all.
Nothing really catastrophic happens (no infinite loops, no runtimes, no crashes... that sort of things).

Direct conversion through moles_per_unit doesn't really work. The amount of moles spessmen breathe each time is several orders of magnitude too small for the reagents datum to even notice. So a high multiplier (at this commit, 100) is needed to get the reagent added and processed.
This breaks conservation of matter and I can foresee a bit of a exploit with dyalisis. But mobs are reagent blackholes, so fuck it.
At this commit, you need many thousands of units added to the room to get reagents added in a smallish room like Chemistry. Numbers have to be tweaked.